All requested data successfully scraped
All data is scraped and logged to the console (title, product URL,
image URL and price). Data is in JSON format.

Errors need to be handled and converted to user friendly message.

JSON from scraped data needs to be stored into correctly named CSV file
using the built-in JS Date object and correct parsing of the dates to
create the file name.